fix: rename reset_previous_style to restore_matplotlib_rcparams
